pyproject.toml
[project]
name = "langraph-supervisor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.0.8",
    "langchain-google-genai>=3.1.0",
    "langchain-mcp-adapters>=0.1.13",
    "langgraph-supervisor>=0.0.31",
]
